Create rten-base crate for shared utilities
Add a crate which lives at the root of the dependency tree, containing common
utilities used by other RTen crates.

The crate is initially populated with iterator traits and utilities used by both
rten-tensor and rten:

 - The `SplitIterator` trait and `ParIter` struct used to simplify
   creating parallel iterators.
 - The `range_chunks` and `range_chunks_exact` functions
