+++ rten-base/README.md
@@ -1,0 +1,4 @@
+# rten-base
+
+This crate contains common utilities used by other RTen crates. It sits at the
+base of the dependency tree.
