+++ rten-base/src/lib.rs
@@ -1,0 +1,6 @@
+//! Common utilities used by other RTen crates.
+//!
+//! This is an internal crate which contains various utilities that don't have
+//! a more dedicated home.
+
+pub mod iter;
