+++ rten-tensor/src/copy.rs
@@ -1,6 +1,6 @@
 use std::mem::MaybeUninit;
-use std::ops::Range;
 
+use rten_base::iter::{range_chunks, range_chunks_exact};
 use smallvec::SmallVec;
 
 use crate::assume_init::AssumeInit;
@@ -10,97 +10,6 @@ use crate::{AsView, Layout};
 use crate::{
     Matrix, MatrixLayout, MatrixMut, NdTensorView, NdTensorViewMut, TensorView, TensorViewMut,
 };
-
-/// Iterator returned by [`range_chunks`].
-pub struct RangeChunks {
-    remainder: Range<usize>,
-    chunk_size: usize,
-}
-
-impl Iterator for RangeChunks {
-    type Item = Range<usize>;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        if !self.remainder.is_empty() {
-            let start = self.remainder.start;
-            let end = (start + self.chunk_size).min(self.remainder.end);
-            self.remainder.start += self.chunk_size;
-            Some(start..end)
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let len = self.remainder.len().div_ceil(self.chunk_size);
-        (len, Some(len))
-    }
-}
-
-impl ExactSizeIterator for RangeChunks {}
-
-impl std::iter::FusedIterator for RangeChunks {}
-
-/// Return an iterator over sub-ranges of `range`. If `range.len()` is not a
-/// multiple of `chunk_size` then the final chunk will be shorter.
-#[inline]
-pub fn range_chunks(range: Range<usize>, chunk_size: usize) -> RangeChunks {
-    RangeChunks {
-        remainder: range,
-        chunk_size,
-    }
-}
-
-pub struct RangeChunksExact {
-    remainder: Range<usize>,
-    chunk_size: usize,
-}
-
-impl RangeChunksExact {
-    /// Return the part of the range that has not yet been visited.
-    #[inline]
-    pub fn remainder(&self) -> Range<usize> {
-        self.remainder.clone()
-    }
-}
-
-impl Iterator for RangeChunksExact {
-    type Item = Range<usize>;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.remainder.len() >= self.chunk_size {
-            let start = self.remainder.start;
-            let end = start + self.chunk_size;
-            self.remainder.start += self.chunk_size;
-            Some(start..end)
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let len = self.remainder.len() / self.chunk_size;
-        (len, Some(len))
-    }
-}
-
-impl ExactSizeIterator for RangeChunksExact {}
-
-impl std::iter::FusedIterator for RangeChunksExact {}
-
-/// Return an iterator over sub-ranges of `range`. If `range.len()` is not a
-/// multiple of `chunk_size` then there will be a remainder after iteration
-/// completes, available via [`RangeChunksExact::remainder`].
-pub fn range_chunks_exact(range: Range<usize>, chunk_size: usize) -> RangeChunksExact {
-    RangeChunksExact {
-        remainder: range,
-        chunk_size,
-    }
-}
 
 /// Tile size for blocked copy. A tile should fit in registers for 32-bit
 /// values.
